♻️ Move happening registration to single page
Jævlig mye cleanere.
